refactor(webapp): extract shared TemplateDropzone component
The import page's drop area is now a component, so other screens can
reuse it instead of copying the render prop and its styles.
